+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * chore(deps): patch `@nuxt/vite-builder`
 * docs(nuxt.config): reduce component meta bundle size
 * fix(module): detect lazy components when using `experimental.componentDetection`
+* fix(NavigationMenu/Tabs): ensure proper badge display
 
 ## [2.0.6](https://github.com/bitrix24/b24ui/compare/v2.0.5...v2.0.6) (2025-10-28)